change packets to be more intuitive

+++ src/types.cpp
@@ -105,7 +105,7 @@ void writeString(int sockfd, std::string string) {
     auto cstr = string.c_str();
     writeVarInt(sockfd, strlen(cstr));
     //std::cout << strlen(cstr) << ": " << cstr << std::endl;
-    if (send(sockfd, cstr, strlen(cstr), 0) < strlen(cstr)) {
+    if (send(sockfd, cstr, strlen(cstr), 0) < (long int)strlen(cstr)) {
         exit(1);
     } 
 }
